feat(tezos_tokens): c++ review fixes
